+++ .dev-doc/main/CHANGELOG.md
@@ -33,3 +33,7 @@
 - 03:55 style: Refine Dev Flow Overview hierarchy and remove redundant path control
 - 03:58 chore: 初始化项目脚手架
 - 04:26 fix: ISSUE-I005：Test TASK-T040 fail:running 5 tests
+
+## 2026-08-10
+- 00:27 fix: ISSUE-I006：Test TASK-T054 fail:running 4 tests
+- 00:54 fix: ISSUE-I007：Test TASK-T054 fail:running 4 tests